refactor: test/lotto/model로 이동
