+++ src/test/java/lotto/model/WinningNumberTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
